lots of improvements, including idea pager fix. Feed cards improvements, and correct data initialization.

+++ app/src/main/java/com/example/track/data/implementations/ideas/SavingsCardRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.example.track.data.implementations.ideas
 
-import com.example.track.data.other.converters.convertLocalDateToDate
 import com.example.track.data.database.incomeRelated.IncomeDao
 import com.example.track.data.implementations.incomes.IncomesCategoriesListRepositoryImpl
+import com.example.track.data.other.converters.convertLocalDateToDate
 import com.example.track.domain.models.idea.Idea
 import com.example.track.domain.models.incomes.IncomeItem
 import com.example.track.domain.repository.ideas.SavingsCardRepository
@@ -14,8 +14,6 @@ class SavingsCardRepositoryImpl(
     private val incomesCategoriesListRepositoryImpl: IncomesCategoriesListRepositoryImpl
 ) : SavingsCardRepository {
     override suspend fun addToSavings(idea: Idea, value: Float, isIncludedInBudget: Boolean) {
-       // if (idea.relatedToAllCategories == true) { //relatedToAllCategories
-            //ideaDao.update(idea.copy(currentValue = idea.currentValue?.plus(value)))
             incomeDao.insert(
                 IncomeItem(
                     currencyTicker = "sd",
@@ -31,11 +29,11 @@ class SavingsCardRepositoryImpl(
 
 
     override fun requestPlannedSavings(idea: Idea): Float {
-        return idea.goal.toFloat()
+        return idea.goal
     }
 
     override fun requestIncludedInBudget(idea: Idea): Boolean? {
-        return false//idea.relatedToAllCategories
+        return false  //idea.relatedToAllCategories
     }
 
     override fun requestCompletenessValue(idea: Idea): Float {
